refactor: 서버 코드 리팩토링

+++ routes/users.js
@@ -1,50 +1,46 @@
 import { MemberStatus } from '@prisma/client';
 import express from 'express';
-import prisma from '../prisma/prismaClient.js'; // prisma를 가져옵니다.
+import prisma from '../prisma/prismaClient.js';
 import { authenticateToken } from '../utils/authMiddleware.js';
 import {
   generateAccessToken,
   generateRefreshToken,
 } from '../utils/jwtUtils.js';
 
-var router = express.Router();
+const router = express.Router();
 
 router.post('/', async (req, res, next) => {
   const { email, authProvider } = req.body;
 
-  let user = await prisma.user.findUnique({
-    where: { email: email, authProvider: authProvider },
-  });
-
-  if (user && user.memberStatus === MemberStatus.Delete) {
-    await prisma.routine.deleteMany({
-      where: { userId: user.id },
+  try {
+    let user = await prisma.user.findUnique({
+      where: { email, authProvider },
     });
 
-    await prisma.user.delete({
-      where: { id: user.id },
-    });
+    if (user && user.memberStatus === MemberStatus.Delete) {
+      await prisma.routine.deleteMany({ where: { userId: user.id } });
+      await prisma.user.delete({ where: { id: user.id } });
+      user = null;
+    }
 
-    user = null;
+    if (!user) {
+      user = await prisma.user.create({
+        data: {
+          email,
+          authProvider,
+          memberStatus: MemberStatus.Register,
+        },
+      });
+    }
+
+    const tokens = {
+      accessToken: generateAccessToken(user),
+      refreshToken: generateRefreshToken(user),
+    };
+    res.json(tokens);
+  } catch (error) {
+    next(error);
   }
-
-  if (!user) {
-    user = await prisma.user.create({
-      data: {
-        email: email,
-        authProvider: authProvider,
-        memberStatus: MemberStatus.Register,
-      },
-    });
-  }
-
-  const accessToken = generateAccessToken(user);
-  const refreshToken = generateRefreshToken(user);
-
-  return res.json({
-    accessToken: accessToken,
-    refreshToken: refreshToken,
-  });
 });
 
 router.get('/', authenticateToken, (req, res, next) => {
@@ -62,18 +58,11 @@ router.put('/', authenticateToken, async (req, res) => {
   const { name, age, job, challenges, gender } = req.body;
 
   try {
-    const user = await prisma.user.update({
+    const updatedUser = await prisma.user.update({
       where: { id: req.user.id },
-      data: {
-        name,
-        age,
-        job,
-        challenges,
-        gender,
-      },
+      data: { name, age, job, challenges, gender },
     });
-
-    res.json(user);
+    res.json(updatedUser);
   } catch (error) {
     res
       .status(500)
@@ -83,15 +72,14 @@ router.put('/', authenticateToken, async (req, res) => {
 
 router.put('/withdraw', authenticateToken, async (req, res) => {
   try {
-    const user = await prisma.user.update({
-      where: { id: req.user?.id },
+    const deletedUser = await prisma.user.update({
+      where: { id: req.user.id },
       data: {
         memberStatus: MemberStatus.Delete,
         deletedAt: new Date(),
       },
     });
-
-    res.json(user);
+    res.json(deletedUser);
   } catch (error) {
     res
       .status(500)
